modules/website/backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==1.10.18
email-validator==1.3.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
alembic==1.12.1
